requirements.txt
pillow>=9.0.0
simplekml>=1.3.0
exifread>=3.0.0
piexif>=1.1.3
